change loader cache to checksum cache
