+++ move-vm/integration-tests/src/tests/bad_storage_tests.rs
@@ -512,7 +512,7 @@ struct BogusStorage {
 impl ModuleResolver for BogusStorage {
     type Error = PartialVMError;
 
-    fn get_module_checksum(&self, _module_id: &ModuleId) -> Result<Option<[u8; 32]>, Self::Error> {
+    fn get_checksum(&self, _module_id: &ModuleId) -> Result<Option<[u8; 32]>, Self::Error> {
         Err(PartialVMError::new(self.bad_status_code))
     }
