+++ move-vm/runtime/Cargo.toml
@@ -22,6 +22,7 @@ sha3 = "0.9.1"
 smallbitvec = "2.5.1"
 tracing = "0.1.26"
 triomphe = "0.1.9"
+lru = "0.12"
 
 move-bytecode-verifier = { path = "../../move-bytecode-verifier" }
 move-core-types = { path = "../../move-core/types" }
