+++ move-vm/runtime/src/loader/resolver.rs
@@ -1,0 +1,351 @@
+use std::sync::Arc;
+
+use move_binary_format::{
+    access::{ModuleAccess, ScriptAccess},
+    errors::{PartialVMError, PartialVMResult},
+    file_format::{
+        Constant, ConstantPoolIndex, FieldHandleIndex, FieldInstantiationIndex,
+        FunctionHandleIndex, FunctionInstantiationIndex, SignatureIndex,
+        StructDefInstantiationIndex, StructDefinitionIndex,
+    },
+};
+use move_core_types::{value::MoveTypeLayout, vm_status::StatusCode};
+use move_vm_types::loaded_data::runtime_types::{AbilityInfo, StructType, Type};
+
+use super::{
+    function::Function, loader::MAX_TYPE_INSTANTIATION_NODES, module::Module, script::Script,
+    Loader,
+};
+
+// A simple wrapper for a `Module` or a `Script` in the `Resolver`
+enum BinaryType {
+    Module(Arc<Module>),
+    Script(Arc<Script>),
+}
+
+// A Resolver is a simple and small structure allocated on the stack and used by the
+// interpreter. It's the only API known to the interpreter and it's tailored to the interpreter
+// needs.
+pub(crate) struct Resolver<'a> {
+    loader: &'a Loader,
+    binary: BinaryType,
+}
+
+impl<'a> Resolver<'a> {
+    pub(crate) fn for_module(loader: &'a Loader, module: Arc<Module>) -> Self {
+        let binary = BinaryType::Module(module);
+        Self { loader, binary }
+    }
+
+    pub(crate) fn for_script(loader: &'a Loader, script: Arc<Script>) -> Self {
+        let binary = BinaryType::Script(script);
+        Self { loader, binary }
+    }
+
+    //
+    // Constant resolution
+    //
+
+    pub(crate) fn constant_at(&self, idx: ConstantPoolIndex) -> &Constant {
+        match &self.binary {
+            BinaryType::Module(module) => module.module.constant_at(idx),
+            BinaryType::Script(script) => script.script.constant_at(idx),
+        }
+    }
+
+    //
+    // Function resolution
+    //
+
+    pub(crate) fn function_from_handle(
+        &self,
+        idx: FunctionHandleIndex,
+    ) -> PartialVMResult<Arc<Function>> {
+        let idx = match &self.binary {
+            BinaryType::Module(module) => module.function_at(idx.0),
+            BinaryType::Script(script) => script.function_at(idx.0),
+        };
+        self.loader.function_at(idx)
+    }
+
+    pub(crate) fn function_from_instantiation(
+        &self,
+        idx: FunctionInstantiationIndex,
+    ) -> PartialVMResult<Arc<Function>> {
+        let func_inst = match &self.binary {
+            BinaryType::Module(module) => module.function_instantiation_at(idx.0),
+            BinaryType::Script(script) => script.function_instantiation_at(idx.0),
+        };
+        self.loader.function_at(&func_inst.handle)
+    }
+
+    pub(crate) fn instantiate_generic_function(
+        &self,
+        idx: FunctionInstantiationIndex,
+        type_params: &[Type],
+    ) -> PartialVMResult<Vec<Type>> {
+        let func_inst = match &self.binary {
+            BinaryType::Module(module) => module.function_instantiation_at(idx.0),
+            BinaryType::Script(script) => script.function_instantiation_at(idx.0),
+        };
+        let mut instantiation = vec![];
+        for ty in &func_inst.instantiation {
+            instantiation.push(self.subst(ty, type_params)?);
+        }
+        // Check if the function instantiation over all generics is larger
+        // than MAX_TYPE_INSTANTIATION_NODES.
+        let mut sum_nodes = 1u64;
+        for ty in type_params.iter().chain(instantiation.iter()) {
+            sum_nodes = sum_nodes.saturating_add(self.loader.count_type_nodes(ty));
+            if sum_nodes > MAX_TYPE_INSTANTIATION_NODES {
+                return Err(PartialVMError::new(StatusCode::TOO_MANY_TYPE_NODES));
+            }
+        }
+        Ok(instantiation)
+    }
+
+    #[allow(unused)]
+    pub(crate) fn type_params_count(&self, idx: FunctionInstantiationIndex) -> usize {
+        let func_inst = match &self.binary {
+            BinaryType::Module(module) => module.function_instantiation_at(idx.0),
+            BinaryType::Script(script) => script.function_instantiation_at(idx.0),
+        };
+        func_inst.instantiation.len()
+    }
+
+    //
+    // Type resolution
+    //
+
+    pub(crate) fn get_struct_type(&self, idx: StructDefinitionIndex) -> PartialVMResult<Type> {
+        let struct_def = match &self.binary {
+            BinaryType::Module(module) => module.struct_at(idx),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        };
+        Ok(Type::Struct {
+            id: struct_def.id.clone(),
+            ability: AbilityInfo::struct_(struct_def.abilities),
+        })
+    }
+
+    pub(crate) fn get_struct_type_generic(
+        &self,
+        idx: StructDefInstantiationIndex,
+        ty_args: &[Type],
+    ) -> PartialVMResult<Type> {
+        let struct_inst = match &self.binary {
+            BinaryType::Module(module) => module.struct_instantiation_at(idx.0),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        };
+
+        // Before instantiating the type, count the # of nodes of all type arguments plus
+        // existing type instantiation.
+        // If that number is larger than MAX_TYPE_INSTANTIATION_NODES, refuse to construct this type.
+        // This prevents constructing larger and larger types via struct instantiation.
+        let mut sum_nodes = 1u64;
+        for ty in ty_args.iter().chain(struct_inst.instantiation.iter()) {
+            sum_nodes = sum_nodes.saturating_add(self.loader.count_type_nodes(ty));
+            if sum_nodes > MAX_TYPE_INSTANTIATION_NODES {
+                return Err(PartialVMError::new(StatusCode::TOO_MANY_TYPE_NODES));
+            }
+        }
+
+        let struct_ = &struct_inst.definition_struct_type;
+        Ok(Type::StructInstantiation {
+            id: struct_.id.clone(),
+            ty_args: triomphe::Arc::new(
+                struct_inst
+                    .instantiation
+                    .iter()
+                    .map(|ty| self.subst(ty, ty_args))
+                    .collect::<PartialVMResult<_>>()?,
+            ),
+            ability: AbilityInfo::generic_struct(
+                struct_.abilities,
+                struct_.phantom_ty_args_mask.clone(),
+            ),
+        })
+    }
+
+    pub(crate) fn get_field_type(&self, idx: FieldHandleIndex) -> PartialVMResult<Type> {
+        match &self.binary {
+            BinaryType::Module(module) => {
+                let handle = &module.field_handles[idx.0 as usize];
+
+                Ok(handle.definition_struct_type.fields[handle.offset].clone())
+            }
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        }
+    }
+
+    pub(crate) fn get_field_type_generic(
+        &self,
+        idx: FieldInstantiationIndex,
+        ty_args: &[Type],
+    ) -> PartialVMResult<Type> {
+        let field_instantiation = match &self.binary {
+            BinaryType::Module(module) => &module.field_instantiations[idx.0 as usize],
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        };
+
+        let instantiation_types = field_instantiation
+            .instantiation
+            .iter()
+            .map(|inst_ty| inst_ty.subst(ty_args))
+            .collect::<PartialVMResult<Vec<_>>>()?;
+        field_instantiation.definition_struct_type.fields[field_instantiation.offset]
+            .subst(&instantiation_types)
+    }
+
+    pub(crate) fn get_struct_fields(
+        &self,
+        idx: StructDefinitionIndex,
+    ) -> PartialVMResult<Arc<StructType>> {
+        match &self.binary {
+            BinaryType::Module(module) => Ok(module.struct_at(idx)),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        }
+    }
+
+    pub(crate) fn instantiate_generic_struct_fields(
+        &self,
+        idx: StructDefInstantiationIndex,
+        ty_args: &[Type],
+    ) -> PartialVMResult<Vec<Type>> {
+        let struct_inst = match &self.binary {
+            BinaryType::Module(module) => module.struct_instantiation_at(idx.0),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        };
+        let struct_type = &struct_inst.definition_struct_type;
+
+        let instantiation_types = struct_inst
+            .instantiation
+            .iter()
+            .map(|inst_ty| inst_ty.subst(ty_args))
+            .collect::<PartialVMResult<Vec<_>>>()?;
+        struct_type
+            .fields
+            .iter()
+            .map(|ty| ty.subst(&instantiation_types))
+            .collect::<PartialVMResult<Vec<_>>>()
+    }
+
+    fn single_type_at(&self, idx: SignatureIndex) -> &Type {
+        match &self.binary {
+            BinaryType::Module(module) => module.single_type_at(idx),
+            BinaryType::Script(script) => script.single_type_at(idx),
+        }
+    }
+
+    pub(crate) fn instantiate_single_type(
+        &self,
+        idx: SignatureIndex,
+        ty_args: &[Type],
+    ) -> PartialVMResult<Type> {
+        let ty = self.single_type_at(idx);
+        if !ty_args.is_empty() {
+            self.subst(ty, ty_args)
+        } else {
+            Ok(ty.clone())
+        }
+    }
+
+    pub(crate) fn subst(&self, ty: &Type, ty_args: &[Type]) -> PartialVMResult<Type> {
+        self.loader.subst(ty, ty_args)
+    }
+
+    //
+    // Fields resolution
+    //
+
+    pub(crate) fn field_offset(&self, idx: FieldHandleIndex) -> usize {
+        match &self.binary {
+            BinaryType::Module(module) => module.field_offset(idx),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have field instructions"),
+        }
+    }
+
+    pub(crate) fn field_instantiation_offset(&self, idx: FieldInstantiationIndex) -> usize {
+        match &self.binary {
+            BinaryType::Module(module) => module.field_instantiation_offset(idx),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have field instructions"),
+        }
+    }
+
+    pub(crate) fn field_count(&self, idx: StructDefinitionIndex) -> u16 {
+        match &self.binary {
+            BinaryType::Module(module) => module.field_count(idx.0),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        }
+    }
+
+    pub(crate) fn field_instantiation_count(&self, idx: StructDefInstantiationIndex) -> u16 {
+        match &self.binary {
+            BinaryType::Module(module) => module.field_instantiation_count(idx.0),
+            BinaryType::Script(_) => unreachable!("Scripts cannot have type instructions"),
+        }
+    }
+
+    pub(crate) fn field_handle_to_struct(&self, idx: FieldHandleIndex) -> PartialVMResult<Type> {
+        match &self.binary {
+            BinaryType::Module(module) => {
+                let struct_ = &module.field_handles[idx.0 as usize].definition_struct_type;
+                Ok(Type::Struct {
+                    id: struct_.id.clone(),
+                    ability: AbilityInfo::struct_(struct_.abilities),
+                })
+            }
+            BinaryType::Script(_) => unreachable!("Scripts cannot have field instructions"),
+        }
+    }
+
+    pub(crate) fn field_instantiation_to_struct(
+        &self,
+        idx: FieldInstantiationIndex,
+        args: &[Type],
+    ) -> PartialVMResult<Type> {
+        match &self.binary {
+            BinaryType::Module(module) => {
+                let struct_ = &module.field_instantiations[idx.0 as usize].definition_struct_type;
+                Ok(Type::StructInstantiation {
+                    id: struct_.id.clone(),
+                    ty_args: triomphe::Arc::new(
+                        module.field_instantiations[idx.0 as usize]
+                            .instantiation
+                            .iter()
+                            .map(|ty| ty.subst(args))
+                            .collect::<PartialVMResult<Vec<_>>>()?,
+                    ),
+                    ability: AbilityInfo::generic_struct(
+                        struct_.abilities,
+                        struct_.phantom_ty_args_mask.clone(),
+                    ),
+                })
+            }
+            BinaryType::Script(_) => unreachable!("Scripts cannot have field instructions"),
+        }
+    }
+
+    pub(crate) fn type_to_type_layout(&self, ty: &Type) -> PartialVMResult<MoveTypeLayout> {
+        self.loader.type_to_type_layout(ty)
+    }
+
+    pub(crate) fn type_to_type_layout_with_identifier_mappings(
+        &self,
+        ty: &Type,
+    ) -> PartialVMResult<(MoveTypeLayout, bool)> {
+        self.loader.type_to_type_layout_with_identifier_mappings(ty)
+    }
+
+    pub(crate) fn type_to_fully_annotated_layout(
+        &self,
+        ty: &Type,
+    ) -> PartialVMResult<MoveTypeLayout> {
+        self.loader.type_to_fully_annotated_layout(ty)
+    }
+
+    // get the loader
+    pub(crate) fn loader(&self) -> &Loader {
+        self.loader
+    }
+}
