+++ move-vm/runtime/src/loader/type_loader.rs
@@ -4,14 +4,14 @@
 use move_binary_format::{
     binary_views::BinaryIndexedView, errors::PartialVMResult, file_format::SignatureToken,
 };
-use move_vm_types::loaded_data::runtime_types::{AbilityInfo, StructNameIndex, Type};
+use move_vm_types::loaded_data::runtime_types::{AbilityInfo, StructIdentifier, Type};
 use triomphe::Arc as TriompheArc;
 
 // `intern_type` converts a signature token into the in memory type representation used by the MoveVM.
 pub fn intern_type(
     module: BinaryIndexedView,
     tok: &SignatureToken,
-    struct_name_table: &[StructNameIndex],
+    struct_identifier: &[StructIdentifier],
 ) -> PartialVMResult<Type> {
     let res = match tok {
         SignatureToken::Bool => Type::Bool,
@@ -25,32 +25,32 @@ pub fn intern_type(
         SignatureToken::Signer => Type::Signer,
         SignatureToken::TypeParameter(idx) => Type::TyParam(*idx),
         SignatureToken::Vector(inner_tok) => {
-            let inner_type = intern_type(module, inner_tok, struct_name_table)?;
+            let inner_type = intern_type(module, inner_tok, struct_identifier)?;
             Type::Vector(TriompheArc::new(inner_type))
-        },
+        }
         SignatureToken::Reference(inner_tok) => {
-            let inner_type = intern_type(module, inner_tok, struct_name_table)?;
+            let inner_type = intern_type(module, inner_tok, struct_identifier)?;
             Type::Reference(Box::new(inner_type))
-        },
+        }
         SignatureToken::MutableReference(inner_tok) => {
-            let inner_type = intern_type(module, inner_tok, struct_name_table)?;
+            let inner_type = intern_type(module, inner_tok, struct_identifier)?;
             Type::MutableReference(Box::new(inner_type))
-        },
+        }
         SignatureToken::Struct(sh_idx) => {
             let struct_handle = module.struct_handle_at(*sh_idx);
             Type::Struct {
-                idx: struct_name_table[sh_idx.0 as usize],
+                id: struct_identifier[sh_idx.0 as usize].clone(),
                 ability: AbilityInfo::struct_(struct_handle.abilities),
             }
-        },
+        }
         SignatureToken::StructInstantiation(sh_idx, tys) => {
             let type_args: Vec<_> = tys
                 .iter()
-                .map(|tok| intern_type(module, tok, struct_name_table))
+                .map(|tok| intern_type(module, tok, struct_identifier))
                 .collect::<PartialVMResult<_>>()?;
             let struct_handle = module.struct_handle_at(*sh_idx);
             Type::StructInstantiation {
-                idx: struct_name_table[sh_idx.0 as usize],
+                id: struct_identifier[sh_idx.0 as usize].clone(),
                 ty_args: TriompheArc::new(type_args),
                 ability: AbilityInfo::generic_struct(
                     struct_handle.abilities,
@@ -61,7 +61,7 @@ pub fn intern_type(
                         .collect(),
                 ),
             }
-        },
+        }
     };
     Ok(res)
 }
