+++ move-vm/runtime/src/move_vm.rs
@@ -3,24 +3,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    config::VMConfig,
-    data_cache::TransactionDataCache,
-    loader::Loader,
-    native_extensions::NativeContextExtensions,
-    native_functions::{NativeFunction, NativeFunctions},
-    runtime::VMRuntime,
-    session::Session,
+    config::VMConfig, data_cache::TransactionDataCache, native_extensions::NativeContextExtensions,
+    native_functions::NativeFunction, runtime::VMRuntime, session::Session,
 };
 use move_binary_format::errors::{Location, PartialVMError, VMResult};
 use move_core_types::{
-    account_address::AccountAddress, identifier::Identifier, resolver::MoveResolver,
+    account_address::AccountAddress, identifier::Identifier, resolver::MoveResolver, language_storage::TypeTag, value::MoveTypeLayout,
 };
 
-#[derive(Clone)]
 pub struct MoveVM {
     pub(crate) runtime: VMRuntime,
-    pub(crate) natives: NativeFunctions,
-    pub(crate) vm_config: VMConfig,
 }
 
 impl MoveVM {
@@ -35,10 +27,8 @@ impl MoveVM {
         vm_config: VMConfig,
     ) -> VMResult<Self> {
         Ok(Self {
-            runtime: VMRuntime::new(),
-            natives: NativeFunctions::new(natives)
+            runtime: VMRuntime::new(natives, vm_config)
                 .map_err(|err| err.finish(Location::Undefined))?,
-            vm_config,
         })
     }
 
@@ -73,7 +63,24 @@ impl MoveVM {
             move_vm: self,
             data_cache: TransactionDataCache::new(remote),
             native_extensions,
-            loader: Loader::new(self.natives.clone(), self.vm_config.clone()),
         }
+    }
+
+    pub fn get_fully_annotated_type_layout(
+        &self,
+        data_cache: &TransactionDataCache,
+        type_tag: &TypeTag,
+    ) -> VMResult<MoveTypeLayout> {
+        self.runtime
+            .loader
+            .get_fully_annotated_type_layout(type_tag, data_cache)
+    }
+
+    pub fn flush_unused_module_cache(&mut self) {
+        self.runtime.loader.flush_unused_module_cache()
+    }
+
+    pub fn flush_unused_script_cache(&mut self) {
+        self.runtime.loader.flush_unused_script_cache()
     }
 }
