+++ move-vm/runtime/src/session.rs
@@ -3,10 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    config::VMConfig,
-    data_cache::TransactionDataCache,
-    loader::{LoadedFunction, Loader},
-    move_vm::MoveVM,
+    config::VMConfig, data_cache::TransactionDataCache, loader::LoadedFunction, move_vm::MoveVM,
     native_extensions::NativeContextExtensions,
 };
 use bytes::Bytes;
@@ -27,13 +24,12 @@ use move_core_types::{
 };
 use move_vm_types::{
     gas::GasMeter,
-    loaded_data::runtime_types::{StructNameIndex, StructType, Type},
+    loaded_data::runtime_types::{StructIdentifier, StructType, Type},
     values::{GlobalValue, Value},
 };
 use std::{borrow::Borrow, sync::Arc};
 
 pub struct Session<'r, 'l> {
-    pub(crate) loader: Loader,
     pub(crate) move_vm: &'l MoveVM,
     pub(crate) data_cache: TransactionDataCache<'r>,
     pub(crate) native_extensions: NativeContextExtensions<'r>,
@@ -90,7 +86,7 @@ impl<'r, 'l> Session<'r, 'l> {
             function_name,
             ty_args,
             args,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             &mut self.native_extensions,
@@ -113,7 +109,7 @@ impl<'r, 'l> Session<'r, 'l> {
             function_name,
             ty_args,
             args,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             &mut self.native_extensions,
@@ -132,7 +128,7 @@ impl<'r, 'l> Session<'r, 'l> {
             func,
             instantiation,
             args,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             &mut self.native_extensions,
@@ -167,7 +163,7 @@ impl<'r, 'l> Session<'r, 'l> {
             script,
             ty_args,
             args,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             &mut self.native_extensions,
@@ -220,7 +216,7 @@ impl<'r, 'l> Session<'r, 'l> {
         self.move_vm.runtime.publish_module_bundle(
             modules,
             sender,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             Compatibility::full_check(),
@@ -238,7 +234,7 @@ impl<'r, 'l> Session<'r, 'l> {
         self.move_vm.runtime.publish_module_bundle(
             modules,
             sender,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             compat_config,
@@ -254,7 +250,7 @@ impl<'r, 'l> Session<'r, 'l> {
         self.move_vm.runtime.publish_module_bundle(
             modules,
             sender,
-            &self.loader,
+            &self.move_vm.runtime.loader,
             &mut self.data_cache,
             gas_meter,
             Compatibility::no_check(),
@@ -272,7 +268,7 @@ impl<'r, 'l> Session<'r, 'l> {
     /// This MUST NOT be called if there is a previous invocation that failed with an invariant violation.
     pub fn finish(self) -> VMResult<ChangeSet> {
         self.data_cache
-            .into_effects(&self.loader)
+            .into_effects(&self.move_vm.runtime.loader)
             .map_err(|e| e.finish(Location::Undefined))
     }
 
@@ -281,7 +277,7 @@ impl<'r, 'l> Session<'r, 'l> {
         resource_converter: &dyn Fn(Value, MoveTypeLayout, bool) -> PartialVMResult<Resource>,
     ) -> VMResult<Changes<Bytes, [u8; 32], Resource>> {
         self.data_cache
-            .into_custom_effects(resource_converter, &self.loader)
+            .into_custom_effects(resource_converter, &self.move_vm.runtime.loader)
             .map_err(|e| e.finish(Location::Undefined))
     }
 
@@ -293,7 +289,7 @@ impl<'r, 'l> Session<'r, 'l> {
             ..
         } = self;
         let change_set = data_cache
-            .into_effects(&self.loader)
+            .into_effects(&self.move_vm.runtime.loader)
             .map_err(|e| e.finish(Location::Undefined))?;
         Ok((change_set, native_extensions))
     }
@@ -311,25 +307,9 @@ impl<'r, 'l> Session<'r, 'l> {
             ..
         } = self;
         let change_set = data_cache
-            .into_custom_effects(resource_converter, &self.loader)
+            .into_custom_effects(resource_converter, &self.move_vm.runtime.loader)
             .map_err(|e| e.finish(Location::Undefined))?;
         Ok((change_set, native_extensions))
-    }
-
-    /// Same like `finish`, but also extracts the native context extensions from the session.
-    pub fn finish_with_extensions_with_loader(
-        self,
-    ) -> VMResult<(ChangeSet, NativeContextExtensions<'r>, Loader)> {
-        let Session {
-            data_cache,
-            native_extensions,
-            loader,
-            ..
-        } = self;
-        let change_set = data_cache
-            .into_effects(&loader)
-            .map_err(|e| e.finish(Location::Undefined))?;
-        Ok((change_set, native_extensions, loader))
     }
 
     /// Try to load a resource from remote storage and create a corresponding GlobalValue
@@ -339,7 +319,8 @@ impl<'r, 'l> Session<'r, 'l> {
         addr: AccountAddress,
         ty: &Type,
     ) -> PartialVMResult<(&mut GlobalValue, Option<NumBytes>)> {
-        self.data_cache.load_resource(&self.loader, addr, ty)
+        self.data_cache
+            .load_resource(&self.move_vm.runtime.loader, addr, ty)
     }
 
     /// Get the serialized format of a `CompiledModule` given a `ModuleId`.
@@ -361,7 +342,9 @@ impl<'r, 'l> Session<'r, 'l> {
         ty_args: Vec<TypeTag>,
     ) -> VMResult<LoadedFunctionInstantiation> {
         let (_, instantiation) =
-            self.loader
+            self.move_vm
+                .runtime
+                .loader
                 .load_script(script.borrow(), &ty_args, &self.data_cache)?;
         Ok(instantiation)
     }
@@ -373,12 +356,16 @@ impl<'r, 'l> Session<'r, 'l> {
         function_name: &IdentStr,
         expected_return_type: &Type,
     ) -> VMResult<(LoadedFunction, LoadedFunctionInstantiation)> {
-        let (func, instantiation) = self.loader.load_function_with_type_arg_inference(
-            module_id,
-            function_name,
-            expected_return_type,
-            &self.data_cache,
-        )?;
+        let (func, instantiation) = self
+            .move_vm
+            .runtime
+            .loader
+            .load_function_with_type_arg_inference(
+                module_id,
+                function_name,
+                expected_return_type,
+                &self.data_cache,
+            )?;
         Ok((func, instantiation))
     }
 
@@ -389,7 +376,7 @@ impl<'r, 'l> Session<'r, 'l> {
         function_name: &IdentStr,
         type_arguments: &[TypeTag],
     ) -> VMResult<LoadedFunctionInstantiation> {
-        let (_, _, instantiation) = self.loader.load_function(
+        let (_, _, instantiation) = self.move_vm.runtime.loader.load_function(
             module_id,
             function_name,
             type_arguments,
@@ -399,20 +386,23 @@ impl<'r, 'l> Session<'r, 'l> {
     }
 
     pub fn load_type(&self, type_tag: &TypeTag) -> VMResult<Type> {
-        self.loader.load_type(type_tag, &self.data_cache)
+        self.move_vm
+            .runtime
+            .loader
+            .load_type(type_tag, &self.data_cache)
     }
 
     pub fn get_type_layout(&self, type_tag: &TypeTag) -> VMResult<MoveTypeLayout> {
-        self.loader.get_type_layout(type_tag, &self.data_cache)
-    }
-
-    pub fn get_fully_annotated_type_layout(&self, type_tag: &TypeTag) -> VMResult<MoveTypeLayout> {
-        self.loader
-            .get_fully_annotated_type_layout(type_tag, &self.data_cache)
+        self.move_vm
+            .runtime
+            .loader
+            .get_type_layout(type_tag, &self.data_cache)
     }
 
     pub fn get_type_tag(&self, ty: &Type) -> VMResult<TypeTag> {
-        self.loader
+        self.move_vm
+            .runtime
+            .loader
             .type_to_type_tag(ty)
             .map_err(|e| e.finish(Location::Undefined))
     }
@@ -432,16 +422,22 @@ impl<'r, 'l> Session<'r, 'l> {
     }
 
     pub fn get_vm_config(&self) -> &'l VMConfig {
-        &self.move_vm.vm_config
+        &self.move_vm.runtime.loader.vm_config
     }
 
-    pub fn get_struct_type(&self, index: StructNameIndex) -> Option<Arc<StructType>> {
-        self.loader.get_struct_type_by_idx(index).ok()
+    pub fn get_struct_type(&self, id: &StructIdentifier) -> Option<Arc<StructType>> {
+        self.move_vm
+            .runtime
+            .loader
+            .get_struct_type_by_identifier(&id)
+            .ok()
     }
 
     /// Load a module into VM's code cache
     pub fn load_module(&self, module_id: &ModuleId) -> VMResult<Arc<CompiledModule>> {
-        self.loader
+        self.move_vm
+            .runtime
+            .loader
             .load_module(module_id, &self.data_cache)
             .map(|arc_module| arc_module.arc_module())
     }
@@ -459,11 +455,18 @@ impl<'r, 'l> Session<'r, 'l> {
     ///
     /// TODO: in the new loader architecture, as the loader is visible to the adapter, one would
     ///   call this directly via the loader instead of the VM.
-    pub fn with_module_metadata<T, F>(&self, module: &ModuleId, f: F) -> Option<T>
+    pub fn with_module_metadata<T, F>(&self, module: &ModuleId, f: F) -> PartialVMResult<Option<T>>
     where
         F: FnOnce(&[Metadata]) -> Option<T>,
     {
-        f(&self.loader.get_module(module)?.module().metadata)
+        let checksum = self.data_cache.load_checksum(module)?;
+
+        Ok(self
+            .move_vm
+            .runtime
+            .loader
+            .get_module(&checksum)
+            .and_then(|v| f(&v.compiled_module().metadata)))
     }
 }
