+++ move-vm/test-utils/Cargo.toml
@@ -15,7 +15,7 @@ anyhow = "1.0.52"
 bytes = "1.4.0"
 once_cell = "1.7.2"
 serde = { version = "1.0.124", features = ["derive", "rc"] }
-sha3 = "0.9.1"
+sha3 = "0.10"
 
 move-binary-format = { path = "../../move-binary-format" }
 move-core-types = { path = "../../move-core/types" }
