+++ move-vm/types/src/values/values_impl.rs
@@ -2175,7 +2175,7 @@ impl Vector {
             | Type::Vector(_)
             | Type::Struct { .. }
             | Type::StructInstantiation {
-                idx: _, ty_args: _, ..
+                id: _, ty_args: _, ..
             } => Value(ValueImpl::Container(Container::Vec(Rc::new(RefCell::new(
                 elements.into_iter().map(|v| v.0).collect(),
             ))))),
