+++ tools/move-cli/Cargo.toml
@@ -25,6 +25,7 @@ serde_yaml = "0.8.17"
 tempfile = "3.2.0"
 toml_edit =  { version = "0.14.3", features = ["easy"] }
 walkdir = "2.3.1"
+sha3 = "0.10"
 
 bcs = { workspace = true }
 move-binary-format = { path = "../../move-binary-format" }
