+++ tools/move-unit-test/src/test_runner.rs
@@ -345,13 +345,13 @@ impl SharedTestingConfig {
                         Some(ExpectedFailure::Expected) => {
                             output.pass(function_name);
                             stats.test_success(test_run_info, test_plan);
-                        },
+                        }
                         Some(ExpectedFailure::ExpectedWithError(expected_err))
                             if expected_err == &actual_err =>
                         {
                             output.pass(function_name);
                             stats.test_success(test_run_info, test_plan);
-                        },
+                        }
                         Some(ExpectedFailure::ExpectedWithCodeDEPRECATED(code))
                             if actual_err.0 == StatusCode::ABORTED
                                 && actual_err.1.is_some()
@@ -359,7 +359,7 @@ impl SharedTestingConfig {
                         {
                             output.pass(function_name);
                             stats.test_success(test_run_info, test_plan);
-                        },
+                        }
                         // incorrect cases
                         Some(ExpectedFailure::ExpectedWithError(expected_err)) => {
                             output.fail(function_name);
@@ -372,7 +372,7 @@ impl SharedTestingConfig {
                                 ),
                                 test_plan,
                             )
-                        },
+                        }
                         Some(ExpectedFailure::ExpectedWithCodeDEPRECATED(expected_code)) => {
                             output.fail(function_name);
                             stats.test_failure(
@@ -387,7 +387,7 @@ impl SharedTestingConfig {
                                 ),
                                 test_plan,
                             )
-                        },
+                        }
                         None if err.major_status() == StatusCode::OUT_OF_GAS => {
                             // Ran out of ticks, report a test timeout and log a test failure
                             output.timeout(function_name);
@@ -400,7 +400,7 @@ impl SharedTestingConfig {
                                 ),
                                 test_plan,
                             )
-                        },
+                        }
                         None => {
                             output.fail(function_name);
                             stats.test_failure(
@@ -412,9 +412,9 @@ impl SharedTestingConfig {
                                 ),
                                 test_plan,
                             )
-                        },
+                        }
                     }
-                },
+                }
                 Ok(_) => {
                     // Expected the test to fail, but it executed
                     if test_info.expected_failure.is_some() {
@@ -433,7 +433,7 @@ impl SharedTestingConfig {
                         output.pass(function_name);
                         stats.test_success(test_run_info, test_plan);
                     }
-                },
+                }
             }
         }
 
@@ -528,7 +528,7 @@ impl SharedTestingConfig {
                         test_plan,
                     );
                     return stats;
-                },
+                }
             };
 
             let (res, duration) = self.execute_via_evm(&yul_code);
@@ -570,7 +570,7 @@ impl SharedTestingConfig {
                         ),
                         test_plan,
                     );
-                },
+                }
 
                 // Test expected to succeed, but aborted.
                 (None, ExitReason::Revert(_)) => {
@@ -588,7 +588,7 @@ impl SharedTestingConfig {
                         ),
                         test_plan,
                     )
-                },
+                }
 
                 // Expect the test to abort with a specific code.
                 (
@@ -621,7 +621,7 @@ impl SharedTestingConfig {
                             test_plan,
                         );
                     }
-                },
+                }
 
                 // Test expected to abort but succeeded.
                 (
@@ -637,18 +637,18 @@ impl SharedTestingConfig {
                         TestFailure::new(FailureReason::no_error(), test_run_info(), None, None),
                         test_plan,
                     )
-                },
+                }
 
                 // Test succeeded or failed as expected.
                 (None, ExitReason::Succeed(_))
                 | (Some(ExpectedFailure::Expected), ExitReason::Revert(_)) => {
                     output.pass(function_name);
                     stats.test_success(test_run_info(), test_plan);
-                },
+                }
 
                 (exp, reason) => {
                     unreachable!("Unexpected (exp, exit reason) pair: ({:?}, {:?}). This should not have happened.", exp, reason)
-                },
+                }
             }
         }
